local_services/costing_engine/pyproject.toml
[project]
name = "costing-engine"
version = "0.1.0"
description = "Costing Engine Workflow Service"
requires-python = ">=3.14"
dependencies = [
    "sqladmin",
    "itsdangerous",
    "fastapi[standard]>=0.115.0",
    "uvicorn[standard]>=0.30.0",
    "sqlmodel>=0.0.21",
    "asyncpg>=0.29.0",
    "alembic>=1.13.0",
    "python-jose[cryptography]>=3.3.0",
    "httpx>=0.27.0",
    "sf-hamilton[async]>=1.49.0",
    "python-dotenv>=1.0.1",
]

[dependency-groups]
dev = [
    "pytest>=8.3.0",
    "pytest-asyncio>=0.23.0",
]
